newline stomper and fixed warnings
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@473 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/MapReader.cpp
@@ -118,7 +118,7 @@ void CMapReader::UnpackTerrain(CFileUnpacker& unpacker)
 	// unpack tile data
 	u32 tilesPerSide=m_MapSize*PATCH_SIZE;
 	m_Tiles.resize(SQR(tilesPerSide));
-	unpacker.UnpackRaw(&m_Tiles[0],sizeof(STileDesc)*m_Tiles.size());	
+	unpacker.UnpackRaw(&m_Tiles[0],(u32)(sizeof(STileDesc)*m_Tiles.size()));	
 }
 
 // ApplyData: take all the input data, and rebuild the scene from it
